fix(plans): move fixture out of __tests__/ to unblock Vercel build
Vercel's Next.js production build excludes `__tests__/` from the
compile graph. Since `page.tsx` does a dev-only `await import("./...
__tests__/fixtures/plan-scenarios")` for the `?scenario=` override,
TypeScript still resolves the import at build time → module not
found → build fails.

Move:
- app/admin/support/plans/__tests__/fixtures/plan-scenarios.ts
  → app/admin/support/plans/_fixtures/plan-scenarios.ts

Update import paths in:
- app/admin/support/plans/page.tsx (dev override)
- scripts/screenshot-plan-scenarios.ts
- scripts/capture-plan-scenarios.ts
- docs/features/provider-plan-sdk-alignment/architecture.md (file index)

Underscore prefix on `_fixtures/` signals "feature-private, not a
route" without triggering Vercel's __tests__/ exclusion. The fixture
content is unchanged.
